# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AuthResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiresIn" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tokenType" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserResponse"
          }
        },
        "required" : [ "accessToken", "expiresIn", "tokenType", "user" ],
        "type" : "object"
      },
...
      "DriverProfileRequest" : {
        "properties" : {
          "licenseNumber" : {
            "minLength" : 1,
            "pattern" : "^[A-Za-z0-9-]{6,40}$",
            "type" : "string"
          },
          "vehicle" : {
            "$ref" : "#/components/schemas/VehicleRequest"
          }
        },
        "required" : [ "licenseNumber", "vehicle" ],
        "type" : "object"
      },
      "DriverResponse" : {
        "properties" : {
          "availability" : {
            "enum" : [ "OFFLINE", "AVAILABLE", "ON_TRIP" ],
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "licenseNumber" : {
            "type" : "string"
          },
          "ratingAvg" : {
            "type" : [ "number", "null" ]
          },
          "ratingCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "rejectionReason" : {
            "type" : [ "string", "null" ]
          },
          "vehicle" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/VehicleResponse"
            }, {
              "type" : "null"
            } ]
          },
          "verificationStatus" : {
            "enum" : [ "PENDING", "VERIFIED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          },
          "verifiedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "availability", "createdAt", "email", "fullName", "id", "licenseNumber", "ratingCount", "verificationStatus" ],
        "type" : "object"
      },
...
      "GeoPoint" : {
        "properties" : {
          "lat" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          },
          "lng" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        },
        "required" : [ "lat", "lng" ],
        "type" : "object"
      },
      "GoOnlineRequest" : {
        "properties" : {
          "location" : {
            "$ref" : "#/components/schemas/GeoPoint"
          }
        },
        "required" : [ "location" ],
        "type" : "object"
      },
      "LocationUpdateRequest" : {
        "properties" : {
          "accuracyMeters" : {
            "format" : "double",
            "maximum" : 10000.0,
            "minimum" : 0.0,
            "type" : [ "number", "null" ]
          },
          "headingDeg" : {
            "format" : "int32",
            "maximum" : 359,
            "minimum" : 0,
            "type" : [ "integer", "null" ]
          },
          "location" : {
            "$ref" : "#/components/schemas/GeoPoint"
          },
          "recordedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "speedMps" : {
            "format" : "double",
            "maximum" : 100.0,
            "minimum" : 0.0,
            "type" : [ "number", "null" ]
          }
        },
        "required" : [ "location", "recordedAt" ],
        "type" : "object"
      },
      "LoginRequest" : {
        "properties" : {
          "email" : {
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
...
      "NearbyDriverResponse" : {
        "properties" : {
          "category" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          },
          "distanceMeters" : {
            "format" : "double",
            "type" : [ "number", "null" ]
          },
          "driverId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "position" : {
            "$ref" : "#/components/schemas/GeoPoint"
          }
        },
        "required" : [ "category", "position" ],
        "type" : "object"
      },
...
      "RegisterRequest" : {
        "properties" : {
          "accountType" : {
            "enum" : [ "PASSENGER", "DRIVER" ],
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "fullName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "phone" : {
            "pattern" : "^\\+[1-9]\\d{7,14}$",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "accountType", "email", "fullName", "password" ],
        "type" : "object"
      },
...
      "UserResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "phone" : {
            "type" : [ "string", "null" ]
          },
          "role" : {
            "enum" : [ "PASSENGER", "DRIVER", "ADMIN" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "SUSPENDED" ],
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "email", "fullName", "id", "role", "status" ],
        "type" : "object"
      },
...
      "VehicleRequest" : {
        "properties" : {
          "category" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          },
          "color" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "make" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "model" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "modelYear" : {
            "format" : "int32",
            "minimum" : 1990,
            "type" : "integer"
          },
          "plateNumber" : {
            "minLength" : 1,
            "pattern" : "^[A-Za-z0-9 -]{4,20}$",
            "type" : "string"
          },
          "seats" : {
            "format" : "int32",
            "maximum" : 8,
            "minimum" : 2,
            "type" : "integer"
          }
        },
        "required" : [ "category", "color", "make", "model", "modelYear", "plateNumber", "seats" ],
        "type" : "object"
      },
      "VehicleResponse" : {
        "properties" : {
          "category" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "make" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "modelYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "plateNumber" : {
            "type" : "string"
          },
          "seats" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "category", "color", "id", "make", "model", "modelYear", "plateNumber", "seats" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Real-time ride-hailing platform API. Conventions: docs/api.md",
    "title" : "RideFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ ],
        "summary" : "Sign in; returns an access token and sets the refresh-token cookie",
        "tags" : [ "Auth" ]
      }
    },
...
    "/api/auth/register" : {
      "post" : {
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ ],
        "summary" : "Create a passenger or driver account",
        "tags" : [ "Auth" ]
      }
    },
    "/api/drivers/location" : {
      "post" : {
        "operationId" : "reportLocation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LocationUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        },
        "summary" : "Report the current GPS position (REST fallback for the WebSocket stream)",
        "tags" : [ "Drivers" ]
      }
    },
...
    "/api/drivers/me/profile" : {
      "post" : {
        "operationId" : "submitProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DriverProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Submit licence and vehicle for verification",
        "tags" : [ "Drivers" ]
      }
    },
    "/api/drivers/me/vehicle" : {
      "put" : {
        "operationId" : "replaceVehicle",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VehicleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Replace the active vehicle; only while offline",
        "tags" : [ "Drivers" ]
      }
    },
    "/api/drivers/nearby" : {
      "get" : {
        "operationId" : "nearby",
        "parameters" : [ {
          "in" : "query",
          "name" : "lat",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "lng",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "radiusMeters",
          "required" : false,
          "schema" : {
            "default" : 3000,
            "format" : "int32",
            "maximum" : 10000,
            "minimum" : 100,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "category",
          "required" : false,
          "schema" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/NearbyDriverResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Available cars near a point (coarsened and anonymous for passengers)",
        "tags" : [ "Drivers" ]
      }
    },
...
    "/api/drivers/online" : {
      "post" : {
        "operationId" : "goOnline",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GoOnlineRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Go online at the given position",
        "tags" : [ "Drivers" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ]
}